pyproject.toml
[project]
name = "banana-eye"
version = "0.1.0"
description = "FastAPI server for generating aerial views using Vertex AI and Mapbox"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "google-generativeai>=0.8.0",
    "requests>=2.31.0",
    "pillow>=10.0.0",
    "pydantic>=2.5.0",
    "python-multipart>=0.0.6",
    "python-dotenv>=1.0.0",
    "google-genai>=1.32.0",
]
